[C] Reorganize hooks dir
Use camelCase for subfolder names; flattened folder structure
