+++ client/src/hooks/index.js
@@ -1,37 +1,37 @@
-export { default as useDispatchAnnotations } from "./annotations/use-dispatch-annotations";
-export { default as useSelectAnnotations } from "./annotations/use-select-annotations";
+export { default as useDispatchAnnotations } from "./useDispatchAnnotations";
+export { default as useSelectAnnotations } from "./useSelectAnnotations";
 
-export { default as useDispatchMyCollected } from "./me/use-dispatch-my-collected";
-export { default as useSelectMyCollected } from "./me/use-select-my-collected";
-export { default as useDispatchMyAnnotatedTexts } from "./me/use-dispatch-my-annotated-texts";
-export { default as useSelectMyAnnotatedTexts } from "./me/use-select-my-annotated-texts";
-export { default as useDispatchMyReadingGroups } from "./me/use-dispatch-my-reading-groups";
-export { default as useSelectMyReadingGroups } from "./me/use-select-my-reading-groups";
+export { default as useDispatchMyCollected } from "./useDispatchMyCollected";
+export { default as useSelectMyCollected } from "./useSelectMyCollected";
+export { default as useDispatchMyAnnotatedTexts } from "./useDispatchMyAnnotatedTexts";
+export { default as useSelectMyAnnotatedTexts } from "./useSelectMyAnnotatedTexts";
+export { default as useDispatchMyReadingGroups } from "./useDispatchMyReadingGroups";
+export { default as useSelectMyReadingGroups } from "./useSelectMyReadingGroups";
 
-export { default as useDispatchReadingGroupMembers } from "./reading-group-memberships/use-dispatch-reading-group-members";
-export { default as useSelectReadingGroupMembers } from "./reading-group-memberships/use-select-reading-group-members";
-export { default as useDispatchReadingGroupMembership } from "./reading-group-memberships/use-dispatch-reading-group-membership";
-export { default as useGrabReadingGroupMembership } from "./reading-group-memberships/use-grab-reading-group-membership";
-export { default as useSelectReadingGroupMembership } from "./reading-group-memberships/use-select-reading-group-membership";
+export { default as useDispatchReadingGroupMembers } from "./useDispatchReadingGroupMembers";
+export { default as useSelectReadingGroupMembers } from "./useSelectReadingGroupMembers";
+export { default as useDispatchReadingGroupMembership } from "./useDispatchReadingGroupMembership";
+export { default as useGrabReadingGroupMembership } from "./useGrabReadingGroupMembership";
+export { default as useSelectReadingGroupMembership } from "./useSelectReadingGroupMembership";
 
-export { default as useDispatchPublicReadingGroups } from "./reading-groups/use-dispatch-public-reading-groups";
-export { default as useSelectPublicReadingGroups } from "./reading-groups/use-select-public-reading-groups";
-export { default as useDispatchReadingGroup } from "./reading-groups/use-dispatch-reading-group";
-export { default as useSelectReadingGroup } from "./reading-groups/use-select-reading-group";
-export { default as useDispatchReadingGroupCategories } from "./reading-groups/use-dispatch-reading-group-categories";
-export { default as useSelectReadingGroupCategories } from "./reading-groups/use-select-reading-group-categories";
-export { default as useDispatchReadingGroupCollected } from "./reading-groups/use-dispatch-reading-group-collected";
-export { default as useSelectReadingGroupCollected } from "./reading-groups/use-select-reading-group-collected";
+export { default as useDispatchPublicReadingGroups } from "./useDispatchPublicReadingGroups";
+export { default as useSelectPublicReadingGroups } from "./useSelectPublicReadingGroups";
+export { default as useDispatchReadingGroup } from "./useDispatchReadingGroup";
+export { default as useSelectReadingGroup } from "./useSelectReadingGroup";
+export { default as useDispatchReadingGroupCategories } from "./useDispatchReadingGroupCategories";
+export { default as useSelectReadingGroupCategories } from "./useSelectReadingGroupCategories";
+export { default as useDispatchReadingGroupCollected } from "./useDispatchReadingGroupCollected";
+export { default as useSelectReadingGroupCollected } from "./useSelectReadingGroupCollected";
 
-export { default as useDispatchSearchResults } from "./search/use-dispatch-search-results";
-export { default as useSelectSearchResults } from "./search/use-select-search-results";
+export { default as useDispatchSearchResults } from "./useDispatchSearchResults";
+export { default as useSelectSearchResults } from "./useSelectSearchResults";
 
-export { default as useGrabCurrentText } from "./texts/use-grab-current-text";
-export { default as useGrabCurrentTextSection } from "./texts/use-grab-current-text-section";
+export { default as useGrabCurrentText } from "./useGrabCurrentText";
+export { default as useGrabCurrentTextSection } from "./useGrabCurrentTextSection";
 
-export { default as useDeprecatedFetchData } from "./api/use-deprecated-fetchdata";
-export { default as useApiCallback } from "./api/use-api-callback";
-export { default as useFetch } from "./api/use-fetch";
-export { default as usePaginationState } from "./api/use-pagination-state";
+export { default as useDeprecatedFetchData } from "./api/useDeprecatedFetchData";
+export { default as useApiCallback } from "./api/useApiCallback";
+export { default as useFetch } from "./api/useFetch";
+export { default as usePaginationState } from "./api/usePaginationState";
 
-export { default as useCurrentUser } from "./user/use-current-user";
+export { default as useCurrentUser } from "./useCurrentUser";
