+++ client/src/hooks/useDispatchSearchResults/index.js
