+++ client/src/hooks/useGrabCurrentText/index.js
