+++ client/src/hooks/useGrabCurrentTextSection/index.js
@@ -1,7 +1,7 @@
 import { useSelector } from "react-redux";
 import { grab } from "utils/entityUtils";
 
-export default function useGrabCurrentText(match) {
+export default function useGrabCurrentTextSection(match) {
   return useSelector(state =>
     grab("textSections", match.params.sectionId, state.entityStore)
   );
