+++ client/src/hooks/useSelectSearchResults/index.js
@@ -2,7 +2,7 @@ import { useSelector } from "react-redux";
 import { requests } from "api";
 import { select, meta } from "utils/entityUtils";
 
-export default function useSelectReadingGroupAnnotations() {
+export default function useSelectSearchResults() {
   const results = useSelector(state =>
     select(requests.rSearchResults, state.entityStore)
   );
