--- client/src/hooks/window-size/index.js
@@ -1,1 +1,0 @@
-export { default as useWindowSize } from "./use-debounced-window-size";
